bug: Fixing a couple places where overflow or truncation could occur
Fixing a couple places where truncation or overflow could occur based off of code ql analysis

+++ utils/C/openSeaChest/openSeaChest_SMART.c
@@ -344,7 +344,7 @@ int32_t main(int argc, char *argv[])
             }
             else if (strncmp(longopts[optionIndex].name, SCSI_DEFECTS_LONG_OPT_STRING, M_Min(strlen(longopts[optionIndex].name), strlen(SCSI_DEFECTS_LONG_OPT_STRING))) == 0)
             {
-                uint8_t counter = 0;
+                size_t counter = 0;
                 SCSI_DEFECTS_FLAG = true;
                 while (counter < strlen(optarg))
                 {
